Split loan / vault implementation
