+++ src/Makefile.am
@@ -167,11 +167,12 @@ DEFI_CORE_H = \
   masternodes/mn_checks.h \
   masternodes/mn_rpc.h \
   masternodes/res.h \
-  masternodes/tokens.h \
+  masternodes/oracles.h \
   masternodes/poolpairs.h \
+  masternodes/tokens.h \
   masternodes/undo.h \
   masternodes/undos.h \
-  masternodes/oracles.h \
+  masternodes/vault.h \
   memusage.h \
   merkleblock.h \
   miner.h \
@@ -385,10 +386,12 @@ libdefi_server_a_SOURCES = \
   masternodes/rpc_oracles.cpp \
   masternodes/rpc_poolpair.cpp \
   masternodes/rpc_tokens.cpp \
+  masternodes/rpc_vault.cpp \
   masternodes/tokens.cpp \
   masternodes/poolpairs.cpp \
   masternodes/skipped_txs.cpp \
   masternodes/undos.cpp \
+  masternodes/vault.cpp \
   miner.cpp \
   net.cpp \
   net_processing.cpp \
